+++ includes/service-providers/mailchimp/class-newspack-newsletters-mailchimp.php
@@ -870,7 +870,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 			return;
 		}
 		$post = get_post( $post_id );
-		if ( Newspack_Newsletters::NEWSPACK_NEWSLETTERS_CPT !== $post->post_type ) {
+		if ( ! Newspack_Newsletters_Editor::is_editing_email( $post_id ) ) {
 			return;
 		}
 		if ( 'trash' === $post->post_status ) {
@@ -1204,7 +1204,12 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 		}
 
 		foreach ( $by_list as $list_id => $sublists ) {
-			$results[] = $this->add_contact( $contact, $list_id, $sublists );
+			$result = $this->add_contact( $contact, $list_id, $sublists );
+			if ( is_wp_error( $result ) ) {
+				return $result;
+			}
+
+			$results[] = $result;
 		}
 		return $results;
 	}